requirements.txt
aiohttp
Brotli
al-utils-almirai
urllib3
beautifulsoup4
colorama
lxml
pyyaml
jsonschema
